services/requirements.txt
# LLM service (OpenAI-compatible API)
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
